refactor the login controller
